Move stopAllBridges to a controller without a Room
